add all methods for GenerateExcelEquipmentDowntimeReport

+++ src/Belaz.WeldingApp.FileApi.Domain/Dtos/EquipmentDowntimeDto.cs
@@ -4,5 +4,5 @@ public class EquipmentDowntimeDto
 {
     public string Reason { get; set; } = null!;
 
-    public int Time { get; set; } = 0;
+    public double Time { get; set; } = 0;
 }
